--- src/orders/components/OrderRefundProducts/index.ts
@@ -1,2 +1,0 @@
-export * from "./OrderRefundProducts";
-export { default } from "./OrderRefundProducts";
